- Alter database data.

+++ GnemesBackEnd/src/main/kotlin/com/malygos/gnemes/dto/MemePostDto.kt
@@ -2,7 +2,7 @@ package com.malygos.gnemes.dto
 
 data class MemePostCreationDto(
         var tag: List<String>?,
-        var difficulty:Int,
+        var difficulty:Byte,
         var oLSentences: List<String>?,
         var sLSentences: List<String>?,
         var phrase: List<String>?
